refactor(tests): reorganize test files for Section 2 consistency + fix ValidationError kwargs bug
- Rename test_ReAct.py → test_react.py, test_ReAct_extended.py → test_react_extended.py (lowercase)
- Remove trailing underscore from test_oai_.py, test_anthropic_.py, test_nvidia_nim_.py
- Rename *_coverage_boost.py → *_edge_cases.py (behavior-oriented names)
- Rename *_coverage.py → *_edge_cases.py (behavior-oriented names)
- Move test_pile_coverage.py → tests/protocols/generic/test_pile_edge_cases.py
- Split test_field_operable_coverage.py → tests/models/test_field_model_edge_cases.py + test_operable_model_edge_cases.py
- Split test_processor_progression_coverage.py → test_progression_edge_cases.py + test_processor_edge_cases.py
- Split test_react_processor_coverage.py → test_react_processor.py (operations) + test_processor_react_edges.py (generic)
- Fix ValidationError constructor in field_model.py:validate() — pass extra args via details= dict
