+++ lionagi/models/field_model.py
@@ -657,9 +657,11 @@ class FieldModel(Params):
                             )
                             raise ValidationError(
                                 f"Validation failed for {validator_name}",
-                                field_name=field_name,
-                                value=value,
-                                validator_name=validator_name,
+                                details={
+                                    "field_name": field_name,
+                                    "value": value,
+                                    "validator_name": validator_name,
+                                },
                             ) from None
                     except Exception:
                         # If validator raises any other exception, let it propagate
