+++ tests/libs/test_schema_loader_edge_cases.py
