+++ tests/protocols/generic/test_pile_edge_cases.py
