+++ tests/protocols/graph/test_node_edge_cases.py
