+++ tests/service/connections/providers/test_anthropic.py
