+++ tests/service/connections/providers/test_nvidia_nim.py
